Remove derivation paths logic from view model
